Moved dtype_to_init_repr to a _utils module
