+++ py-polars/polars/datatypes/__init__.py
@@ -54,7 +54,6 @@ from polars.datatypes.constructor import (
 )
 from polars.datatypes.convert import (
     dtype_to_ffiname,
-    dtype_to_init_repr,
     dtype_to_py_type,
     is_polars_dtype,
     maybe_cast,
@@ -128,7 +127,6 @@ __all__ = [
     # convert
     "dtype_to_ffiname",
     "dtype_to_py_type",
-    "dtype_to_init_repr",
     "is_polars_dtype",
     "maybe_cast",
     "numpy_char_code_to_dtype",
